a2a-with-strands/agents/orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-opentelemetry-distro>=0.16.0",
    "bedrock-agentcore>=1.6.0",
    "fastapi>=0.135.3",
    "httpx>=0.27.0",
    "strands-agents[a2a,otel]>=1.35.0",
]
